chromeos-bootimage: Compress assets before building
With unified builds we plan to build multiple images, one for each model.
Since they all use the same assets, we only need to compress them once.
So adjust the code to compress the assets before calling build_images().

Print a message while doing it since it takes a while.

BUG=chromium:685670
TEST=emerge-reef chromeos-bootimage
$ find /build/reef/firmware/ -name "image*"
/build/reef/firmware/image.net.bin
/build/reef/firmware/image.bin
/build/reef/firmware/image.dev.bin
/build/reef/firmware/image.serial.bin

Change-Id: I770279b413ca6730f69983352489192a540df0ac
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434059
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>