c50-update: protect from USB failures
The mechanism to prevent rolling reboots in case of cr50 update
failures is flawed: the code witch increments the reboot counter value
is way below several spots where an error can happen which would cause
a reboot. Incidentally a recent change exposed this problem by running
usb_updater in the beginning of the script, this caused rolling
reboots on a device with Suzy-Q connected.

This change makes sure that the counter is saved the first thing. For
that the calculation of the cache file name had to change, since it is
not know at the beginning of the script which image, dev or prod the
H1 is running.

To make things consistent the contents of a list of files is piped
into md5sum, including both cr50 images found in
/opt/google/cr50/firmware and /dev/null just in case the images were
removed (could happen on a debug setup).

Also modified the first usb_updater invocation to always go over
/dev/tmp0 - this would work on all platforms irrespective of servo
connection.

BRANCH=none
BUG=chrome-os-partner:62480
TEST=verified that there is no more rolling reboots when servo is
     connected to eve. Also verified that update to a newer version
     works as expected.

Change-Id: I20fa81c5def4c5ba83bd2d372e04c942d0a7b8e7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/435549
Commit-Ready: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>