media-rules: Fix subsystem name for rk3399 triggers
In CL:433837 we changed the udev-trigger-codec.conf script to only
trigger the events meaningful for the platform the script is run on, but
rk3399 triggers were mistakenly changed to match using --subsystem-match
with "platform", instead of "video4linux", which was incorrect and could
not match devices in question.

Fix the script by making it use the correct subsystem match.

BUG=chromium:682578
BUG=chrome-os-partner:62563
TEST=install on kevin|gru
  open two terminals on DUT
  term 1: udevadm monitor
  term 2: execute changed commands
  expected result: udevadm monitor reports a video4linux uevent for
    each command

Change-Id: I0dd2b2f0dd96a212392d6077b4c07730d2717577
Reviewed-on: https://chromium-review.googlesource.com/435648
Commit-Ready: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>