virtual/vulkan-icd: New ebuild
This package indicates the presence of a Vulkan ICD. We need this ebuild
because libvulkan.so (provided by media-libs/vulkan-loader) does not
provide a Vulkan implementation. It is merely a loader library for the
ICDs, which provide the actual implementation.

We should probably upstream this ebuild to Gentoo.

BUG=chromium:689177
TEST=Confirm that `USE=vulkan emerge-sentry virtual/vulkan-icd`
     installs deps media-libs/vulkan-loader and media-libs/mesa[vulkan].

Change-Id: I29c0c78ee5e016ec5b2c1a08d758382a35ebc567
Reviewed-on: https://chromium-review.googlesource.com/438747
Commit-Ready: Chad Versace <chadversary@chromium.org>
Tested-by: Chad Versace <chadversary@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>