cr50-update: add more logging
cr50-update is causing some issues, but it is hard to tell exactly what
it is doing. This change adds more logging.

This change also moves the usb_update_options definition down to where
usb_update is called. This makes it easier to tell what
usb_update_options is, and we probably want to do as little as possible
before the counter increments.

BUG=chrome-os-partner:62911
BRANCH=none
TEST=/var/log/messages shows more cr50-update messages

Change-Id: Ib5a6010492d1cf10f6100e616722cdcdd2762ada
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/442694
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Scott Collyer <scollyer@chromium.org>