authpolicy: Disable authpolicy on x86
So far, only Samba was disabled on x86, but authpolicy was enabled.
Since authpolicy is unusable without Samba, though, it is cleaner
to disable it. Another reason to disable it is that unit tests are
failing because of a seccomp crash (filters have syscalls not present
on x86).

This change is implemented in chromeos-chrome-9999.ebuild to work
around an issue in the build system and will be moved to
package.use.mask when the ebuild is marked stable.

BUG=chromium:664441
TEST=build_packages for enguarde

Change-Id: Ibc310050aa05a9457432b135f53445d57a215fe3
Reviewed-on: https://chromium-review.googlesource.com/443589
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>