cros-firmware: Add support for unified builds
Add support for creating a firmware update for unified builds. This mostly
relies on the pack_firmware.py script, which is able to read information
from the master configuration to decide what to include in the update for
each model.

BUG=chromium:690573
TEST=USE=unibuild emerge-reef --nodeps chromeos-firmware-reef
With other changes, see that it produces a valid firmware update
Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: Ie88d9afe082b96057b24f42d385185c89defb7b6
Reviewed-on: https://chromium-review.googlesource.com/458900
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>