Allow use of CROS_WORKON_SRCPATH with private overlays
At present we try to access a directory called

   /mnt/host/source/src/private-overlays/src/../chromeos-base/...

when using private-overlays. Adjust the code to use CROS_WORKON_SRCPATH
as a full path so that the ebuild has control over where the source is.
CROS_WORKON_SRCPATH appears not to be used anywhere currently, so it
should be safe to change it.

The CL which exposes this problem is:

https://chrome-internal-review.googlesource.com/c/310315/

BUG=chromium:671391
TEST=With the afore-mentioned CL:
$ emerge-reef --nodeps chromeos-config-bsp
and see that it now completes

Change-Id: Ib8af84b16343ca7798630302ec22dde3698eb47f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/444027