chromeos-accelerometer-init: Check calibration data
Before applying calibration data, check they are sane.
If one of values is incorrect, ignore them all.

Check only once, write the result in a file:
/var/lib/vpd_data_valid

Add unit tests.

BUG=chrome-os-partner:62342
TEST=Unit tests
Check on a machine with bad calibration that the data is ignored.
Check on a machine with good data that the data is used.
Check the values takes ~150ms.
Once checked, an accelerometer is set up within 130ms.

Change-Id: I4d364ff914ffbfffd65965c3a200f037b0b87db8
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/444828
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>