cups: Change USB uri naming scheme to make it friendlier to chrome-matching.
This changes the cups USB uri naming to the form usb://VENDOR_ID/PRODUCT_ID[?serial=SERIAL

where VENDOR_ID and PRODUCT_ID are both 4-digit lowercase hex fields,
and serial is whatever is determined to be the (USB) serial number of
the printer.  This intentionally avoids using any ieee1284 fields in
printer identification, since those fields are subject to fixups and
mangling in cups, so generating the same URI on the chrome side is
fragile.

TEST=manual
BUG=chromium:616866

Change-Id: Ib6bfe3404fedf0c02a9a26b9e1e7db0469d72705
Reviewed-on: https://chromium-review.googlesource.com/445998
Commit-Ready: Justin Carlson <justincarlson@chromium.org>
Tested-by: Justin Carlson <justincarlson@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>