chromeos-accelerometer-init: Set calibration to default when VPD is invalid
If invalid calibration data has been sent to the EC, it may stick until
EC is restarted. In case of invalid VPD data we must set the calibias to
0. We leave calibscale untouched given it is only used for older systems
and has been set to a constant value (1024).

Also, using /var/lib revealed a hidden dependency of accelerometer-init,
chromeos-startup. We need to be sure chromeos-startup has mounted /mnt/stateful
for accelerometer-init to run properly.

BUG=chrome-os-partner:62342
CQ-DEPEND=CL:448158
TEST=Unit tests
On veyron and cave, check by setting vpd_data_valid to "valid" and "invalid"
that the calibbias is changed while the calibscale is unchanged.

Testing on cave, there is a slight decrease in boot time (+25ms to login):
cros-ec-accel initialization is not done while splash screen is being
displayed, but did not add a reverse dependency.

Afer: (on 50 cycles):   Before: (on 44 cycles):
 time  s%     dt  s%	 time  s%     dt  s%  event
 1127  1%  +1127  1%	 1126  2%  +1126  2%  startup
 1552  3%   +425 10%	 1548  5%   +422 15%  startup_done
 2355  3%   +803  6%	 2349  5%   +801  9%  chrome_exec
 2552  4%   +197 41%	 2530  5%   +181 41%  chrome_main
 3988  5%  +1436 14%	 3940  4%  +1410 12%  signin_start
 4004  5%    +16 36%	 3956  4%    +16 33%  signin_wait
 5202  3%  +1198  6%	 5173  2%  +1217  4%  signin_seen
 5347  3%   +145 22%	 5322  2%   +149 14%  login
 7865 12%  +2518 40%	 7788 10%  +2466 33%  network

Change-Id: Ifb77b5c965674b1a855a9506d19e66b831bbed74
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/447882
Reviewed-by: Mike Frysinger <vapier@chromium.org>