cros-firmware: Drop single quotes in the arguments
The _add_param() function currently puts the flag value in single quotes.
This works fine when calling a shell script, but with the Python script
it results in single quotes appearing in the arguments.

There doesn't seem to be any need for this, and with the shell script the
result is the same with double quotes.

Drop the single quote. A future CL will move to using arrays and remove
all this code.

BUG=chromium:690573
TEST=PYTHONPATH=With test CL in place which modifies cros-firmware eclass
to compare the output of pack_firmware.sh and pack_firmware.py:
emerge-reef --nodeps chromeos-firmware-reef

Change-Id: I332ec4a6adb67115ef07e40ec4789c69061f0846
Reviewed-on: https://chromium-review.googlesource.com/447860
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>