python-uinput: patch to build with newer linux-headers
In preparation for linux-headers upgrade, I noticed that python-uinput
stopped to build. A more straightforward way is to upgrade it to a newer
version, but Gentoo upstream is still at 0.10.2 and I am not sure if it
should be upgraded there first.

I am just pulling the patch that fixed the build failure on
linux-headers-4.4 this time.

BUG=b:34841890
TEST=trybot against chromiumos-sdk

Change-Id: Ib66fc666d5b70147a03805ef90f933f085a56d74
Reviewed-on: https://chromium-review.googlesource.com/457918
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>