sys-apps/dbus: upgraded to 1.10.12 and upreved libchrome
Upgraded dbus to Gentoo stable, 1.10.12. Ebuild was tweaked to build
with EAPI5 - had to replace
`--with-systemdsystemunitdir="$(systemd_get_systemunitdir)"' with
`"$(systemd_with_unitdir)"'.

Two patches were dropped as they appear in upstream.
dbus-1.6.8-send-print-fixed.patch was remade.

This dbus upgrade includes a breaking ABI change: public APIs are now
exported in a version specific namespace, i.e., LIBDBUS_1_3, as opposed
to 'Base', which a few chromeos packages assume, including libchrome and
Chrome.

libchrome is easy. We just have to uprev it so it gets re-linked. Chrome
is trickier because most pre-cq/paladin builders are designed to not
build it from source. They use a prebuilt version linked against an
older libdbus. Just so this change gets through pre-cq/cq, we add a
compat hack in the dbus ebuild, that exports the symbols used by Chrome
in the 'Base' name space, in addition to the default LIBDBUS_1_3. Once a
new Chrome is built (which should be linked with the LIBDBUS_1_3 symbols
as they are default), we can remove that hack.

The selected symbols are obtained from `readelf -sW
/build/$BOARD/opt/google/chrome/chrome | grep -i @Base`

For some reason the compat hack doesn't work for the SDK. Since we don't
need it for the SDK anyway, we hide it behind USE=cros_host.

BUG=b:34841890, chromium:704365
TEST=trybots against chromiumos-sdk, x86-mario-release, lakitu-release,
    daisy-release, x86-generic-full, amd64-generic-full,
    caroline-paladin, and caroline-release
CQ-DEPEND=CL:459092, CL:459704, CL:461216, CL:465546

Change-Id: Iee257bb0d459ac20c8871e1d41bdf9d3c87e1c03
Reviewed-on: https://chromium-review.googlesource.com/461717
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>