cros-firmware: Don't unpack files with unified builds
With unified builds file unpacking is done by the pack_firmware.py script
since it is just too complicated to unpack things for all the different
models in the eclass and pass all the resulting filenames for each model
to the script.

Adjust src_unpack() to do nothing when USE=unibuild is active.

Since this is the first CL that uses the unibuild flag in this eclass, add
it to IUSE and bring in the master configuration if needed.

BUG=chromium:690573
TEST=USE=bootimage emerge-reef --nodeps chromeos-firmware-reef

Change-Id: Ie3b0e4f9fb3f72528353bd8fc3182913e781ab4f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/457934
Reviewed-by: Dan Erat <derat@chromium.org>