cros-firmware: Download files for unified builds firmware update
With unified builds the files to be included in the update are listed in
the master configuration. We still need to download them from BCS, so add
a new setup function which handles this.

When unified builds is being used, this new function (called
cros-firmware_setup_source_unibuild()) should be called from the
chromeos-firmware-<board> ebuild instead of cros-firmware_setup_source().

It would be better to have this happen automatically based on the unibuild
USE flag. We could check this in cros-firmware_setup_source() and do the
right thing. However it does not seem to be possible to access USE flags
that early in the eclass.

BUG=chromium:690573
TEST=USE=unibuild emerge-reef --nodeps chromeos-firmware-reef
With other changes, see that it produces a valid firmware update.

Change-Id: Ib76e1adc5a2f72723755a17f3c722ae951436f61
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/458899
Reviewed-by: Mike Frysinger <vapier@chromium.org>