cros-kernel2: support kmemleak via USE
This is useful for detecting memory leaks in the kernel. It's nice to
have an easy USE flag for it.

I tested this works fine on 4.4. It reported "Early log buffer exceeded
(3490), ..." errors when used with the default of 400, so I bump it up
to 16K.

I haven't tested other kernels, but the Kconfig is similar, and
available (but disabled) in the base.config, so the USE flag should work
fine.

BUG=none
TEST=build, boot with 'USE=kmemleak'

Change-Id: I8f9e343d8d2efee6532653c562c972d3fab28016
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/461304
Reviewed-by: Guenter Roeck <groeck@chromium.org>