Add /opt/google/chrome/dbus to include dirs.
This allows D-Bus service files to be copied to the output of the Chrome
build and managed by deploy_chrome instead of requiring an update to an
ebuild file for every new D-Bus service file.

This new file must have a ".conf" extension to be included; D-Bus
ignored an attempted chromium.xml.

BUG=chromium:701617
TEST=Installed package, service files in /opt/google/chrome/dbus work

Change-Id: I0ce8bac696be2a96e1d0ae10d2106bb6918fa47c
Reviewed-on: https://chromium-review.googlesource.com/461330
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>