dev-lang/rust: add rust ebuild to make cross-compiling rustc
This ebuild is based off the portage upstream version, but applies some
patches to so that rustc can target our various toolchains. It also hard
disables eselect, and versioned install directories. The chroot's llvm is not
used because it is incompatible with rust until llvm is updated.

TEST=sudo emerge rust, rustc should be on $PATH
BUG=None

Change-Id: I14c3dd2a2940f5b714cb3c00305aeab5ea510499
Reviewed-on: https://chromium-review.googlesource.com/462276
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>