media-libs/mesa: fix infinite loop when using swrast with vgem disabled
Because of an inverted loop increment, the swrast patch on mesa would
loop forever trying to find the vgem render node if one doesn't exist.
This hang was never encountered before because it only happens on
boards with no vgem AND that use our patched mesa for swrast.

BUG=None
TEST=null_platform_test with no files under /dev/dri

Change-Id: Ide5ecaafd051f93f6261bbf9795902a7e660175c
Reviewed-on: https://chromium-review.googlesource.com/462417
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>