media-libs/mesa: fix the fix for infinite loop when using swrast...
The patch with change-id Ide5ecaafd051f93f6261bbf9795902a7e660175c
patched the swrast path file to not infinite loop by changing the loop's
direction when it should have changed the loops upper bound. This patch
patches the patch file properly, promise.

The testing procedure outline below is strange because the infinite
looping behavior is tricky to trigger on a board. The first line is
intended to confuse mesa into not finding a device for rendering or for
using vgem. The second line instruct null_platform_test to use card0 as
a display and to show all stat/open calls to the render nodes. Before
this pair of patches, one would see a huge list of stat calls to ever
increasing renderD files. After this patch, only renderD128 to
renderD143 should be seen in the calls starting with
'stat("/sys/bus/platform/devices/vgem/drm/'. Note that no screen output
is expected from this test.

TEST=mv /dev/dri /dev/dri2 ;
     strace null_platform_test /dev/dri2/card0 2>&1 | grep renderD
BUG=None

Change-Id: I8f999aa607f6b92d27383324c6087347c2231475
Reviewed-on: https://chromium-review.googlesource.com/463686
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>