add cros-llvm.eclass
We want to add libc++ to default chromiumos-sdk. This requires these
libraries to be installed in cros_setup_toolchain stage.
To do this, we need to set the right toolchain and install prefix for
it. This eclass provides a function to set the right toolchain and
install prefix.

BUG=chromium:620402
TEST=sudo emerge cross-x86_64-cros-linux-gnu/{libunwind, libcxxabi, libcxx}
     x86_64-cros-linux-gnu-clang++ a.o -stdlib=libc++ -lc++ -lc++abi
     sudo CC=clang CXX=clang++ emerge cross-armv7a-cros-linux-gnueabi/libunwind
     usese armv7a-cros-linux-gnueabi-clang to build the package.
     sudo emerge cross-armv7a-cros-linux-gnueabi/libunwind
     usese armv7a-cros-linux-gnueabi-gcc to build the package.
Change-Id: I91f295022b06a963ac86ec6f09fc86d9bdbbbff8
Reviewed-on: https://chromium-review.googlesource.com/466514
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>