cros-unibuild: Allow configuration access without a ROOT
The SRC_URI variable is calculated very early during ebuild processing and
there is no guarantee that there is even a root directory in existence. So
if we want to look up information for this variable we cannot obtain the
configuration in the normal way.

Provide a separate function which can get the configuration directly from
the source ebuild via a symlink.

BUG=chromium:690573
TEST=USE=unibuild emerge-reef --nodeps chromeos-firmware-reef
See that it still manages to set up SRC_URI and download the files

Change-Id: Id355a5456737615518b9b7883cc03ec125f50e2d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/470407
Reviewed-by: Mike Frysinger <vapier@chromium.org>