chromeos-base/factory: Remove virtual/chromeos-interface from dependency.
The chromeos-interface was listed as dependency when the factory program
still depend on Telemetry and chromedriver. We have abandoned that for a
long time and now always building with minimal dependency with the new
Makefile.

It should be safe to remove chromeos-interface to make it easier for
both partners and Google developers to build factory toolkit easier and
faster.

BUG=None
TEST=./setup_board --board eve
     emerge-eve -j 32 factory # success

Change-Id: I88e43cc8281e5f3cde8d2fc138ed4353068e917d
Reviewed-on: https://chromium-review.googlesource.com/471686
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Trybot-Ready: Hung-Te Lin <hungte@chromium.org>