Remove clang syntax checking from valgrind package.
When building with GCC, valgrind uses -mpreferred-stack-boundary=2
which clang does not recognize, so the clang syntax checker fails,
causing the build to fail.  (Not an issue if building directly with
clang, as valgrind checks to see if the main compiler recognizes the
flag or not).

This CL removes clang syntax checking for the valgrind package.

BUG=chromium:710934
TEST=emerge-x86-alex valgrind failed without this change, but passes
with this change.

Change-Id: I14650a112e58042aa6e3f2ed8ec558b302c07f1c
Reviewed-on: https://chromium-review.googlesource.com/475977
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>