Update SDK packages that need to build with GCC.
As part of migrating the host compiler from GCC to LLVM, some
SDK packages do not build with LLVM, so we need to force those
packages to build with GCC, at least for the SDK. This CL does
that.  Once the host compiler is switched, we will investigate
getting these packages to build with clang.

BUG=chromium:674175
TEST=Currently testing with various buildbots.

Change-Id: I1b2d3ef390b775937b186bcb5b09053821cde4b4
Reviewed-on: https://chromium-review.googlesource.com/482475
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>