busybox: Add masked, pristine copy of 1.26.2 ebuild from upstream
This is in preparation for setting "+static" on the ebuild. Upstream had
previously set "+static" which we rely on for our Recovery images.
Now, upstream doesn't set this (though it remains an optional USE flag).
In a follow-up commit, I will set our overlay to "+static" and unmask
this package to finish the upgrade without breaking Recovery.

I followed the instructions at:
http://www.chromium.org/chromium-os/gentoo-package-upgrade-process#TOC-Re-applying-a-patch-after-upgrade

BUG=chromium:710074
TEST=Ran equery + build_packages to confirm that 1.23.1 is still used.

Change-Id: If7864140dcd17be5c38c5bc14f09915f72ba267a
Reviewed-on: https://chromium-review.googlesource.com/485861
Commit-Ready: Jason Clinton <jclinton@chromium.org>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>