Disable USE flag for strongSwan unity plugin
During the 5.5.0 upgrade, the unity plugin got enabled because it
is on by default in new strongSwan builds.  But this plugin has a
known issue that breaks transport mode, resulting in a regression when
interoperating with certain VPN gateways.  Disable the unity plugin
until this is fixed.  We never need Split-Include for transport mode
anyway, and we do not currently support tunnel mode.

BUG=chromium:707139
TEST=manually connect to Sophos UTM test VPN and verify ping
TEST=run autotests: network_VPNConnect.l2tpipsec_{psk,cert,xauth}

Change-Id: Ie317522b3cf4d1f488e3d225d013d305f4ea3c4c
Reviewed-on: https://chromium-review.googlesource.com/489082
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>