biod: add missing dependency on virtual/libfp
biod requires fingerprint support libraries to communicate with the
sensor and to match enrolled templates.
These libraries are relatively closely tied to hardware, so we create a
virtual libfp package that is mostly a placeholder and that will be
overriden by libfp ebuilds coming from the respective board overlays,
similarly to chromeos-bsp.

BUG=b:35585898
TEST=emerge-eve biod
TEST=build_packages --board=gru doesn't break.
CQ-DEPEND=CL:495587

Change-Id: I447be140ddc48e92fdbe241288d499314fa83d6e
Reviewed-on: https://chromium-review.googlesource.com/492506
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>