mesa: fix cros_workon_make error by using default "S" path
The default value of S is "${WORKDIR}/${P}" [1], but the Mesa ebuild
modifies S to capitalize the package name. So instead of
"${WORKDIR}/mesa-9999" the path becomes "${WORKDIR}/Mesa-9999".

It seems that cros_workon_make didn't get the memo; it expects the
lowercase name of the package. This causes "cros_workon_make mesa" to
fail after the first time it is run. It doesn't know that src_prepare
has already been run, so it tries to apply patches that are already
applied to the source tree and fails.

From the commit history it's not clear why this special capitalized
directory name was chosen. Mesa seems to build fine with the default
value of S, and the default value of S lets cros_workon_make work
properly.

Manually uprev'd the stable ebuild since CROS_WORKON_BLACKLIST is set.

BUG=None
TEST=run "cros_workon start mesa", then run "cros_workon_make mesa" twice to verify that rebuilding with no changes succeeds

[1] https://devmanual.gentoo.org/ebuild-writing/variables/

Change-Id: I960acfd6f752de921ddf3f50cccdaaaa1d7ee414
Reviewed-on: https://chromium-review.googlesource.com/491126
Commit-Ready: Nicholas Bishop <nbishop@neverware.com>
Tested-by: Nicholas Bishop <nbishop@neverware.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>