Add U-Boot credits again
This was dropped in an earlier commit. I have verified that the credits do
not appear on snow with the latest release. As suggested, put this in a
new ebuild. It seems to need to install a file to get the license to be
generated, so add a simple README.

See here for earlier discussion:

https://chromium-review.googlesource.com/c/491249/

BUG=b:35515634
TEST=emerge-daisy u-boot-scripts
See that it picks up this ebuild also
Then in chromite/licensing:
./licenses --board daisy -o asc
And see that credits for U-Boot appear in the output file.

Change-Id: Ib3da5177c21e2465c903c7fa4a5071d1505f9bc8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/493831
Reviewed-by: Mike Frysinger <vapier@chromium.org>