Clean up coreboot-utils and change it to be pulled in by userfeedback
sys-apps/coreboot-utils serves two purposes: with cros_host it installs
several utilities that are needed to build a firmware image, and for an
actual board it installs the cbmem utility that is used to query runtime
firmware information (e.g. for chrome://system). (It also installs
cbfstool for a factory use case that should probably be solved by
supplying it externally instead, but that is out of scope for this
bug... see crosbug.com/p/49634.)

However, the dependencies that pull coreboot-utils in are very weird:
the cros_host use case is solved by build-time DEPENDs from the various
firmware ebuilds, which works fine... but on base images it seems that
it only gets installed as an RDEPEND of the depthcharge ebuild that
should've really been a DEPEND anyway. The problem is that the
depthcharge ebuild doesn't always get emerged on system images, because
it depends on the USE=bootimage flag. This is only set for a few boards
that we explicitly selected to continue building ToT firmware for the
CQ, whereas the actual firmware running on Chromebooks comes from
firmware branches. (The chipset-skylake-private overlay also has a
custom dependency on coreboot-utils, presumably to get it to appear on
some boards that didn't have USE=bootimage set on ToT.)

This patch fixes this weirdness by pulling coreboot-utils in from the
userfeedback ebuild instead, which contains the code that actually calls
these utilities. Also add it unconditionally to developer images just in
case. Fix up depthcharge's dependencies and make a few of the installed
utilities conditional to cros_host which aren't really needed outside
the build process.

Removing the sys-apps/pciutils dependency because I have no idea how it
got there in the first place. I don't know of anything in coreboot-utils
that would require it at runtime. If people want it available for
testing purposes, it still gets pulled in directly by chromium-os-test.

Removing the old USE=coreboot flag from all remaining packages while I'm
here... this flag was only used to differentiate coreboot systems from
legacy EFI firmware many years ago, and has become vestigal with recent
firmware build refactorings.

BUG=b:37757150
TEST=None

Change-Id: I576935622128119dafe7660841025ec76e7ec6d6
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/495366