factory_installer: Use installer resource from chromeos-base/factory.
The cros_payload also has to be shared between factory installer and
factory toolkit, so we have to put it into a shared resource for
factory_installer. The resource is now renamed from "cutoff" to
"installer" because it will contain more non-cutoff resource in future.

BUG=chromium:711615
TEST=emerge-eve factory factory_installer
CQ-DEPEND=CL:496227

Change-Id: I03f4d31d7a9a96af4080b2fb966d296673dffb56
Reviewed-on: https://chromium-review.googlesource.com/495948
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>