Add patch for "Allow PolicyLoadStatusSample to override reporting method"
Revision in chromium/src: 6410def39290ed829e0d56b92ab03ddd985a0c04

Adds a patch that cherry-picks a CL required to fix a memory leak in
Chrome OS. PolicyLoadStatusSample reports stats to LinearHistogram
by default, which does not work in Chrome OS. This CL introduces
a base class which just gathers samples without doing anything with
them and can be used in preg_parser.

Adds a patch to the ebuild instead of uprev'ing it because libchrome
has been uprev'ed to the latest and greatest in revision c3f34a3e,
but that change hasn't been picked up by the ebuild yet. The patch
is an intermediate solution. When the ebuild is uprev'ed to the
latest libchrome revision, the change is picked up from a similar
commit to libchrome (CL:497433).

CQ-DEPEND=CL:496128

BUG=chromium:717888
TEST=Compiles.

Change-Id: I8cbac293fc6c24ee97dc514d631612a0debbd162
Reviewed-on: https://chromium-review.googlesource.com/497434
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>