sys-apps/dbus: Avoid traversing /var for pid file and system socket.
This changes the pid file and system socket paths to use /run instead
of /var/run. The latter is an alias for the latter anyways, but
traversing /var is risky since that file system is stateful so can be
messed with by attackers before a verified reboot takes place.

BUG=None
TEST=dbus builds successfully, OS still boots.
CQ-DEPEND=CL:503268

Change-Id: I1548a30d03a46e2b5db02f21a7c3d5bf3c7495f2
Reviewed-on: https://chromium-review.googlesource.com/501927
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>