Temporarily move coreboot-utils dependency back to depthcharge
This is a partial revert of commit 067f826a (Clean up coreboot-utils and
change it to be pulled in by userfeedback). It turns out that installing
coreboot-utils to the base image is currently very deadly to Cyan
boards. While we're fixing that, this patch reverts the coreboot-utils
dependency situation to the (weird and non-sensical) previous state
where it is an RDEPEND of depthcharge. The other parts of that patch are
unrelated and don't need to be reverted.

CQ-DEPEND=CL:502015
BUG=b:37757150,b:38211793,chromium:719786
TEST=None

Change-Id: I3a8b7e7d28fd24ce3a0ecfefcd6cb33f8301592b
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/502094