cros-board: Output the reference board with unified builds
To provide a unified build for an existing board we create a new overlay
called <board>-uni. For example, the unified build for reef is reef-uni.

This is not actually a build target for firmware so adjust
get_current_board_with_variant() to return the original board (here
"reef") in this situation.

This allows us to bring in the reef overlay and build firmware without
creating a special new target in the various firmware ebuilds.

BUG=chromium:718534
TEST=./build_packages --board=reef-uni --nowithautotest --noworkon
Change-Id: Ib3848b0cb1dddfa2fef7024d7928d58a67cc9bb2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/503469
Reviewed-by: Aaron Durbin <adurbin@chromium.org>