cr50: script for background firmware updates
When required, the new script will be invoked at the postinstall phase
of Chrome OS install process to copy the new cr50 image to the H1.

The update would happen only if two conditions are met:

 - the currently running cr50 version is 0.0.19 or older, otherwise
   the system could reset unexpectedly for the user after the update.
   This check is done by trunks_send internally.

 - the resident version of trunks_send is capable of sending the image
   to the H1

If either of these conditions is not true, the update would happen on
the next system restart, under the init scripts' control.

The H1 is guaranteed not to allow fallbacks and to ignore images which
are of the same version as the currently running.

CQ-DEPEND=CL:505251
BUG=b:35580805, chromium:721612
TEST=in concert with CL:49391 verified the following:
    - successful update of cr50 from RW18 to RW19 when updating a reef
      running M57 to M60 (the actual update happens at boot time and
      the system reboots twice)

    - successful update from 60 to 60 with a newer image of cr50, the
      new image was installed in the background, no two restarts were
      required.

Change-Id: Ida5019554d1ac67b2cb74a69819b88c15def3e93
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/505459
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>