eclass: cros-firmware: Create firmware updater script only if available.
Some testing scripts are relying on existence of chromeos-firmwareupdate
to decide if it should extract information from it; the new
chromeos-firmware-null is now always creating an empty updater if image
files are not specified. It seems more reasonable to prevent creating
empty updater instead of changing the test logic in all other test
scripts.

BUG=None
TEST=emerge-$BOARD chromeos-firmware-null

Change-Id: Id8f8fa02f84320600e79116cbad8b4f562f35810
Reviewed-on: https://chromium-review.googlesource.com/505976
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>