chromeos-base/tty: No need to check for existence of crossystem
In reviews of CL:506770 after the CL landed, Mike pointed out that we
don't need to check for the existence of crossystem before running
it.  Remove it.

The original code here was copied from crosh:

  https://chromium.googlesource.com/chromiumos/platform2/+/master/crosh/crosh#116

...but there we're trying to show nice messages.  Here it's not needed
and if crossystem is missing then we'll just fall through to the
'else' clause and stop trying to spawn the agetty.

BUG=chromium:722913
TEST=Hack to call x_crossystem.  No login prompt, no respawns.

Change-Id: I53684d1d6ba097a92f8c80991507560f9eda4808
Reviewed-on: https://chromium-review.googlesource.com/508187
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>